function/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions == 1.11.2
azure-servicebus == 7.8.0